refactor(040): migra packages/shared-data para TypeScript
